[31] Implemented views for statistical reports and filtering

+++ backend/src/model/dto/user_dto.rs
@@ -13,5 +13,5 @@ pub struct UserDTO {
 pub struct UserReportDTO {
     #[serde(flatten)]
     pub user: User,
-    pub participations: usize
+    pub participations: i32
 }
